+++ pyproject.toml
@@ -27,6 +27,9 @@ retry = "==0.9.2"
 geoip2 = "*"
 bcrypt = "*"
 pydantic = "*"
+types-Flask = "^1.1.0"
+types-dataclasses = "^0.1.3"
+types-python-dateutil = "^0.1.2"
 
 [tool.poetry.dev-dependencies]
 pylama = "*"
